requirements.txt
discord.py==2.3.2
openai==1.61.0
python-dotenv==1.0.1
Requests==2.32.3
yt_dlp==2025.1.26
